Code/reports/bug fixes for tutorial (#31)
