+++ arches/medium_core.yaml
@@ -1,7 +1,10 @@
-
 #
 # Set up the pipeline for a 3-wide machine
 #
+
+# Build on top of a small core
+include: small_core.yaml
+
 top.cpu.core0:
   fetch.params.num_to_fetch:   3
   decode.params.num_to_decode: 3
