+++ arches/small_core.yaml
@@ -1,7 +1,7 @@
-
 #
 # Set up the pipeline for a 2-wide machine
 #
+
 top.cpu.core0:
   fetch.params.num_to_fetch:   2
   decode.params.num_to_decode: 2
