+++ layouts/gen_layouts
@@ -1,4 +1,13 @@
 #!/bin/bash
 
-./gen_alf       core.cmd > core.alf
-./gen_alf -t 80 core.cmd > core_80.alf
+# Generate the small core
+./gen_alf -c arch=small       core.cmd > small_core.alf
+./gen_alf -c arch=small -t 80 core.cmd > small_core_80.alf
+
+# Generate the medium core
+./gen_alf       core.cmd > medium_core.alf
+./gen_alf -t 80 core.cmd > medium_core_80.alf
+
+# Generate the big core
+./gen_alf       core.cmd > big_core.alf
+./gen_alf -t 80 core.cmd > big_core_80.alf
