+++ reports/dhry_report.yaml
@@ -1,0 +1,18 @@
+#
+# This is a report definition file for dhrystone, calculating DMIPs
+# per MHz assuming a 1GHz part
+#
+content:
+  top:
+    "cpu.core0.rob.stats.ipc" : "IPC"
+
+    # Assuming 1 iteration is about 265 instructions, calculate the
+    # cycles per iteration
+    "265 / cpu.core0.rob.stats.ipc" : "Cycles Per Iteration"
+
+    # Assume a 1MHz part, determine the iterations per second
+    "1e6 / (265 / cpu.core0.rob.stats.ipc)" : "Iterations Per Second"
+
+    # Assume a 1MHz part, determine the DMIPs based on the VAX 1780
+    # achieving a 1757 Dhrystones per second score
+    "(1e6 / (265 / cpu.core0.rob.stats.ipc)) / 1757" : "DMIPS Per MHz"
