+++ traces/stf_trace_gen/dhrystone/README_C
@@ -1,0 +1,78 @@
+This "shar" file contains the documentation for the
+electronic mail distribution of the Dhrystone benchmark (C version 2.1);
+a companion "shar" file contains the source code.
+(Because of mail length restrictions for some mailers, I have
+split the distribution in two parts.)
+
+For versions in other languages, see the other "shar" files.
+
+Files containing the C version (*.h: Header File, *.c: C Modules)
+
+  dhry.h
+  dhry_1.c
+  dhry_2.c
+  
+The file RATIONALE contains the article 
+
+  "Dhrystone Benchmark: Rationale for Version 2 and Measurement Rules"
+
+which has been published, together with the C source code (Version 2.0),
+in SIGPLAN Notices vol. 23, no. 8 (Aug. 1988), pp. 49-62.
+This article explains all changes that have been made for Version 2,
+compared with the version of the original publication
+in Communications of the ACM vol. 27, no. 10 (Oct. 1984), pp. 1013-1030.
+It also contains "ground rules" for benchmarking with Dhrystone
+which should be followed by everyone who uses the program and publishes
+Dhrystone results.
+
+Compared with the Version 2.0 published in SIGPLAN Notices, Version 2.1
+contains a few corrections that have been made after Version 2.0 was
+distriobuted over the UNIX network Usenet. These small differences between
+Version 2.0 and 2.1 should not affect execution time measurements.
+For those who want to compare the exact contents of both versions,
+the file "dhry_c.dif" contains the differences between the two versions,
+as generated by a file comparison of the corresponding files with the
+UNIX utility "diff".
+
+The file VARIATIONS contains the article
+
+  "Understanding Variations in Dhrystone Performance"
+
+which has been published in Microprocessor Report, May 1989
+(Editor: M. Slater), pp. 16-17. It describes the points that users
+should know if C Dhrystone results are compared.
+
+Recipients of this shar file who perform measurements are asked
+to send measurement results to the author and/or to Rick Richardson.
+Rick Richardson publishes regularly Dhrystone results on the UNIX network
+Usenet. For submissions of results to him (preferably by electronic mail,
+see address in the program header), he has provided a form which is contained
+in the file "submit.frm".
+
+
+The following files are contained in other "shar" files:
+
+Files containing the Ada version (*.s: Specifications, *.b: Bodies):
+
+  d_global.s
+  d_main.b
+  d_pack_1.b
+  d_pack_1.s
+  d_pack_2.b
+  d_pack_2.s
+
+File containing the Pascal version:
+
+  dhry.p
+
+
+February 22, 1990
+
+                 Reinhold P. Weicker
+                 Siemens AG, AUT E 51
+                 Postfach 3220
+                 D-8520 Erlangen
+                 Germany (West)
+
+                 Phone:  [xxx-49]-9131-7-20330  (8-17 Central European Time)
+                 UUCP:   ..!mcsun!unido!estevax!weicker
